refactor(layouts): split FocusedChatWorkstationRail into a directory
Convert the rail into FocusedChatWorkstationRail/ with index.tsx plus
WorkstationSections, WorkstationItemRow, WorkspaceContextRow,
RailItemStatus, railStorage and types. index.tsx re-exports
FocusedChatSessionContext, so the single import site is unchanged.
