+++ src/modules/shared/layouts/FocusedChatWorkstationRail/WorkspaceContextRow.tsx
@@ -1,0 +1,65 @@
+/**
+ * WorkspaceContextRow — a read-only (or click-through) scope row in the rail:
+ * repo, branch, worktree, or the linked work item.
+ */
+import type { LucideIcon } from "lucide-react";
+
+import { WORKSTATION_TRAIL_CONTENT } from "@src/config/workstation/tokens";
+
+export function WorkspaceContextRow({
+  compact = false,
+  icon: Icon,
+  label,
+  onClick,
+  onRequestClose,
+  testId,
+  title,
+}: {
+  compact?: boolean;
+  icon: LucideIcon;
+  label: string;
+  onClick?: () => void;
+  onRequestClose?: () => void;
+  testId?: string;
+  title?: string;
+}) {
+  const className = compact
+    ? "flex h-8 min-w-0 items-center gap-2 overflow-hidden rounded-md px-2 text-text-1"
+    : `${WORKSTATION_TRAIL_CONTENT.row} gap-1.5 overflow-hidden px-2 text-text-1`;
+  const content = (
+    <>
+      <Icon className="shrink-0" size={14} strokeWidth={1.75} />
+      <span
+        className={`min-w-0 flex-1 truncate ${
+          compact ? "text-[13px]" : "text-[12px]"
+        }`}
+      >
+        {label}
+      </span>
+    </>
+  );
+
+  if (onClick) {
+    return (
+      <button
+        type="button"
+        className={`${className} w-full text-left transition-colors hover:bg-fill-2`}
+        title={title ?? label}
+        data-testid={testId}
+        role={compact ? "menuitem" : undefined}
+        onClick={() => {
+          onRequestClose?.();
+          onClick();
+        }}
+      >
+        {content}
+      </button>
+    );
+  }
+
+  return (
+    <div className={className} title={title ?? label} data-testid={testId}>
+      {content}
+    </div>
+  );
+}
